feat: 메뉴바 clear 버튼 공통 hook으로 분리
