+++ app/hooks/index.tsx
@@ -7,3 +7,4 @@ export { useCustomOverlay } from "./useCustomOverlay";
 export { useTargetView } from "./useTargetView";
 export { useImageUpload } from "./useImageUpload";
 export { useMoveTheMap } from "./useMoveTheMap";
+export { useClear } from "./useClear";
